requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.2
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangotoolbox==1.8.0
dlib==19.5.0
drf-nested-routers==0.90.0
face-recognition==1.0.0
face-recognition-models==0.2.0
factory-boy==2.8.1
Faker==0.7.17
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
numpy==1.13.1
olefile==0.44
openapi-codec==1.3.1
Pillow==4.2.1
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
scipy==0.19.1
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
